Changed structure for store related files, added redux-devtools
