+++ package.json
@@ -15,6 +15,7 @@
     "react-router-dom": "^5.2.1",
     "react-scripts": "4.0.3",
     "redux": "^4.1.1",
+    "redux-devtools-extension": "^2.13.9",
     "styled-components": "^5.3.0",
     "web-vitals": "^1.0.1"
   },
