+++ src/components/UserLoginForm/UserLoginForm.js
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 import { Link, withRouter } from 'react-router-dom';
 import { connect } from 'react-redux';
 
-import { authenticate } from '../../actions';
+import { authenticate } from '../../store/actions/auth';
 import { compose } from '../HocUtils/compose';
 
 import { Input } from '../Input/Input';
