+++ src/index.js
@@ -4,7 +4,7 @@ import { AppWithProfileDataAndAuth } from './App';
 import * as serviceWorker from "./serviceWorker";
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import { store } from './store';
+import { store } from './store/store';
 
 ReactDOM.render(
   <React.StrictMode>
