+++ src/store/reducers/profile.js
@@ -1,11 +1,12 @@
-import { STORE_PAYMENT_DATA } from '../actions';
+import { STORE_PAYMENT_DATA, STORE_PAYMENT_ERROR } from '../actions/profile';
 
 const initialState = {
   hasEligiblePayment: false,
   cardName: '',
   cardNumber: '',
   expiryDate: '',
-  cvc: ''
+  cvc: '',
+  error: ''
 };
 
 export function profile (state = initialState, action) {
@@ -22,6 +23,14 @@ export function profile (state = initialState, action) {
         cvc
       }
     }
+    case STORE_PAYMENT_ERROR: {
+      const { error } = action.payload;
+
+      return {
+        ...state,
+        error
+      }
+    }
     default:
       return state
   }
@@ -32,3 +41,4 @@ export const getCardName = state => state.profile.cardName;
 export const getCardNumber = state => state.profile.cardNumber;
 export const getExpiryDate = state => state.profile.expiryDate;
 export const getCvc = state => state.profile.cvc;
+export const getProfileError = state => state.profile.error;
